+++ Libplanet.Tests/Common/Action/Attack.cs
@@ -1,7 +1,5 @@
-using System;
 using System.Collections.Generic;
 using System.Collections.Immutable;
-using System.Text;
 using Libplanet.Action;
 
 namespace Libplanet.Tests.Common.Action
@@ -14,26 +12,29 @@ namespace Libplanet.Tests.Common.Action
             {
                 { "weapon", Weapon },
                 { "target", Target },
+                { "target_address", TargetAddress.ToByteArray() },
             }.ToImmutableDictionary();
 
         public string Weapon { get; set; }
 
         public string Target { get; set; }
 
+        public Address TargetAddress { get; set; }
+
         public override void LoadPlainValue(
             IImmutableDictionary<string, object> plainValue)
         {
             Weapon = (string)plainValue["weapon"];
             Target = (string)plainValue["target"];
+            TargetAddress = new Address((byte[])plainValue["target_address"]);
         }
 
         public override AddressStateMap Execute(IActionContext context)
         {
             var result = new BattleResult();
             AddressStateMap previousStates = context.PreviousStates;
-            Address to = context.To;
 
-            if (previousStates.TryGetValue(to, out object value))
+            if (previousStates.TryGetValue(TargetAddress, out object value))
             {
                 var previousResult = (BattleResult)value;
                 result.UsedWeapons = previousResult.UsedWeapons;
@@ -43,7 +44,7 @@ namespace Libplanet.Tests.Common.Action
             result.UsedWeapons.Add(Weapon);
             result.Targets.Add(Target);
 
-            return (AddressStateMap)previousStates.SetItem(to, result);
+            return (AddressStateMap)previousStates.SetItem(TargetAddress, result);
         }
     }
 }